+++ db/migrate/20210115212305_add_changed_hearing_request_type_to_appeal.rb
@@ -1,0 +1,8 @@
+class AddChangedHearingRequestTypeToAppeal < ActiveRecord::Migration[5.2]
+  def change
+    add_column :appeals,
+               :changed_request_type,
+               :string,
+               comment: "The new hearing type preference for an appellant that needs a hearing scheduled"
+  end
+end
